Реализовано основное задание + доп №1
