+++ data/2.html
@@ -1,8 +1,9 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-	<meta charset="UTF-8">
-</head>
-<body>
-    <h1>Второй файл</h1>
-</body>
+    <head>
+        <meta charset="UTF-8">
+    </head>
+    <body>
+        <h1>Второй файл</h1>
+    </body>
+</html>
